add navbar and reorganize files
