+++ src/config/routes.jsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { Router, Route } from 'react-router'
 
-import Main from '../components/Main'
-import Home from '../components/Home'
-import Profile from '../components/Profile'
-import NotFound from '../components/NotFound'
+import Main from '../components/Main/Main'
+import Home from '../components/Home/Home'
+import Profile from '../components/Profile/Profile'
+import NotFound from '../components/NotFound/NotFound'
 
 var routes = (
   <Router>
