+++ views/index.ejs
@@ -2,7 +2,6 @@
 <html>
   <head>
     <title>Node/React</title>
-    <link rel="stylesheet" href="style.css">
   </head>
   <body>
     <div id="app"><%- content %></div>
